requirements-dev.txt
pytest
ruff
sqlfluff
requests
psycopg2-binary
